[binary_sensor] Use uint8_t for counts, pack bools, add validation
- Use uint8_t for timing_count_ (saves 4 bytes per instance by
  packing with adjacent bool/uint8_t fields, no padding)
- Reorder MultiClickTriggerBase fields to pack uint8_t + 3 bools
  into one 4-byte word
- Add cv.Length(max=255) validation for autorepeat and multi_click
  timing lists to match uint8_t storage

+++ esphome/components/binary_sensor/filter.h
@@ -99,7 +99,7 @@ class AutorepeatFilterBase : public Filter, public Component {
   void next_value_(bool val);
 
   const AutorepeatFilterTiming *timings_{nullptr};
-  size_t timings_count_{0};
+  uint8_t timings_count_{0};
   uint8_t active_timing_{0};
 };
